mgr/dashboard_v2: Refactored `RESTControllerTest`:
* Made `RESTControllerTest` inherent from `ControllerTestCase`
* Refactored `ControllerTestCase`
* Simplified all tests that inherent from  `ControllerTestCase`
