+++ src/pybind/mgr/dashboard_v2/tests/helper.py
@@ -2,41 +2,34 @@
 # pylint: disable=W0212
 from __future__ import absolute_import
 
-import os
-import unittest
+import json
 
-import requests
+import cherrypy
+from cherrypy.test import helper
+
+from ..controllers.auth import Auth
+from ..tools import json_error_page, SessionExpireAtBrowserCloseTool
 
 
-class ControllerTestCase(unittest.TestCase):
-    DASHBOARD_HOST = os.environ.get('DASHBOARD_V2_HOST', "localhost")
-    DASHBOARD_PORT = os.environ.get('DASHBOARD_V2_PORT', 8080)
-
+class ControllerTestCase(helper.CPWebCase):
     def __init__(self, *args, **kwargs):
-        self.dashboard_host = kwargs.pop('dashboard_host') \
-            if 'dashboard_host' in kwargs else self.DASHBOARD_HOST
-        self.dashboard_port = kwargs.pop('dashboard_port') \
-            if 'dashboard_port' in kwargs else self.DASHBOARD_PORT
+        cherrypy.tools.authenticate = cherrypy.Tool('before_handler', Auth.check_auth)
+        cherrypy.tools.session_expire_at_browser_close = SessionExpireAtBrowserCloseTool()
+        cherrypy.config.update({'error_page.default': json_error_page})
         super(ControllerTestCase, self).__init__(*args, **kwargs)
-        self._session = requests.Session()
-        self._resp = None
 
     def _request(self, url, method, data=None):
-        url = "http://{}:{}{}".format(self.dashboard_host, self.dashboard_port,
-                                      url)
-        if method == 'GET':
-            self._resp = self._session.get(url)
-            return self._resp.json()
-        elif method == 'POST':
-            self._resp = self._session.post(url, json=data)
-        elif method == 'DELETE':
-            self._resp = self._session.delete(url, json=data)
-        elif method == 'PUT':
-            self._resp = self._session.put(url, json=data)
-        return None
+        if not data:
+            b = None
+            h = None
+        else:
+            b = json.dumps(data)
+            h = [('Content-Type', 'application/json'),
+                 ('Content-Length', str(len(b)))]
+        self.getPage(url, method=method, body=b, headers=h)
 
     def _get(self, url):
-        return self._request(url, 'GET')
+        self._request(url, 'GET')
 
     def _post(self, url, data=None):
         self._request(url, 'POST', data)
@@ -47,21 +40,15 @@ class ControllerTestCase(unittest.TestCase):
     def _put(self, url, data=None):
         self._request(url, 'PUT', data)
 
-    def cookies(self):
-        return self._resp.cookies
-
     def jsonBody(self):
-        return self._resp.json()
+        body_str = self.body.decode('utf-8') if isinstance(self.body, bytes) else self.body
+        return json.loads(body_str)
 
-    def reset_session(self):
-        self._session = requests.Session()
-
-    def assertJsonBody(self, data):
-        body = self._resp.json()
-        self.assertEqual(body, data)
-
-    def assertBody(self, body):
-        self.assertEqual(self._resp.text, body)
-
-    def assertStatus(self, status):
-        self.assertEqual(self._resp.status_code, status)
+    def assertJsonBody(self, data, msg=None):
+        """Fail if value != self.body."""
+        json_body = self.jsonBody()
+        if data != json_body:
+            if msg is None:
+                msg = 'expected body:\n%r\n\nactual body:\n%r' % (
+                    data, json_body)
+            self._handlewebError(msg)
