+++ src/pybind/mgr/dashboard_v2/tests/test_tools.py
@@ -1,14 +1,12 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
-import json
-
 import cherrypy
 from cherrypy.lib.sessions import RamSession
-from cherrypy.test import helper
 from mock import patch
 
-from ..tools import RESTController, detail_route
+from .helper import ControllerTestCase
+from ..tools import RESTController
 
 
 # pylint: disable=W0613
@@ -50,38 +48,7 @@ class Root(object):
     fooargs = FooArgs()
 
 
-class RESTControllerTest(helper.CPWebCase):
-    def _request(self, url, method, data=None):
-        if not data:
-            b = None
-            h = None
-        else:
-            b = json.dumps(data)
-            h = [('Content-Type', 'application/json'),
-                 ('Content-Length', str(len(b)))]
-        self.getPage(url, method=method, body=b, headers=h)
-
-    def _get(self, url):
-        self._request(url, 'GET')
-
-    def _post(self, url, data=None):
-        self._request(url, 'POST', data)
-
-    def _delete(self, url, data=None):
-        self._request(url, 'DELETE', data)
-
-    def _put(self, url, data=None):
-        self._request(url, 'PUT', data)
-
-    def assertJsonBody(self, data, msg=None):
-        """Fail if value != self.body."""
-        body_str = self.body.decode('utf-8') if isinstance(self.body, bytes) else self.body
-        json_body = json.loads(body_str)
-        if data != json_body:
-            if msg is None:
-                msg = 'expected body:\n%r\n\nactual body:\n%r' % (
-                    data, json_body)
-            self._handlewebError(msg)
+class RESTControllerTest(ControllerTestCase):
 
     @classmethod
     def setup_server(cls):
